refactor: rename context and model folders
