+++ src/pages/Project.tsx
@@ -2,7 +2,7 @@ import { FormEvent, useState } from 'react';
 import { FiDownload, FiPlus, FiX } from 'react-icons/fi';
 import { BiRocket } from 'react-icons/bi';
 
-import { useTask } from '../context/task';
+import { useTask } from '../contexts/task';
 
 import Header from '../components/Header';
 import TaskTimeline from '../components/TaskTimeline';
